requirements.txt
pip~=21.1
bump2version~=1.0.1
wheel~=0.37.1
watchdog~=2.1.9
flake8~=4.0.1
tox~=3.25.1
coverage~=6.4.1
Sphinx~=5.0.2
twine~=4.0.1
httpx~=0.23.0
python-dotenv~=0.20.0
pytest~=7.1.2
